Create rental storage
