+++ create_rental.rb
@@ -22,7 +22,8 @@ def create_rental
   date = gets.chomp
   rental = Rental.new(date, @books[id_book], @people[id_person])
   @rentals << rental
-  # File.write('rentals.json', JSON.pretty_generate(@rentals.map { |r| { date: r.date, book: r.book.title, person: r.person.name } }))
+  File.write('rentals.json', JSON.pretty_generate(@rentals.map { |b| { date: b.date ,person: {name:b.person.name, age: b.person.age, id:b.person.id, type: b.person.class.name}, book:{title:b.book.title, author: b.book.author}  
+  } }))    
   puts 'Rental created successfully'
   start
 end
