+++ people.json
@@ -3,12 +3,18 @@
     "type": "Student",
     "name": "jack",
     "age": 33,
-    "id": 98
+    "id": 232
   },
   {
     "type": "Teacher",
     "name": "jose",
     "age": 33,
-    "id": 778
+    "id": 374
+  },
+  {
+    "type": "Student",
+    "name": "jishan",
+    "age": 23,
+    "id": 610
   }
 ]
